Nuke old SDK
